fully decomp save and combine split save c files
