+++ include/save.h
@@ -3,6 +3,27 @@
 
 #include "time.h"
 
+struct UnkStruct_sub_8011DAC {
+    u8 fill000[0x4];
+    u8 unk004[0x400];
+    u8 unk404[0x10];
+    u32 unk414;
+    u32 unk418;
+    u32 unk41C;
+    u32 unk420;
+    u32 unk424;
+    u32 unk428;
+    u8 fill42C[0x4];
+    u32 unk430;
+    u32 unk434;
+    u32 unk438;
+    u32 unk43C;
+    u32 unk440;
+    u32 unk444;
+    u8 unk448[0x538C];
+};
+
+
 struct UnkStruct_203B184 {
     /* 0x0 */ u32 *unk0;
     /* 0x4 */ u32 *unk4;
@@ -29,6 +50,18 @@ struct UnkStruct_203B184 {
     u32 unk058;
 };
 
+u32 sub_8011C1C(void);
+void sub_8011C28(u32 in);
+s32 sub_8011C34(void);
+void sub_8011C40(s32 in);
+u32 sub_8011C4C(void);
+void sub_8011CA8(u32 *out, s32 size);
+u32 sub_8011DAC(u32 *a);
+u32 sub_8011F9C(s32 *r0, u8 *dest, s32 size);
+bool8 sub_8011FF8(void);
+void sub_8012284(void);
+void sub_8012298(void);
+void sub_80122A8(void);
 void sub_8012468(void);
 u8 sub_8012484(void);
 void sub_8012558(void);
@@ -42,5 +75,16 @@ void sub_8012834(void);
 void sub_8012850(u8 *r0, u32 r1, u8 r2);
 u32 sub_80128B0(void);
 void sub_80129FC(void);
+
+void sub_8012298();
+void sub_80122D0();
+void sub_80122F4();
+void sub_8012300();
+void sub_80122A8();
+
+u32 ReadSaveSector(s32 *a, u8 *dest, s32 size);
+u32 WriteSaveSector(s32 *a, u8 *src, s32 size);
+void CalculateChecksum(u8 *out, u32 size);
+bool8 ValidateChecksum(u8 *in, u32 size);
 
 #endif
