+++ ld_script.txt
@@ -82,8 +82,6 @@ SECTIONS {
         src/debug.o(.text);
         asm/debug.o(.text);
         src/save.o(.text);
-        asm/save.o(.text);
-        src/save_mid.o(.text);
         src/code_8012A18.o(.text);
         src/more_input.o(.text);
         asm/code_8012A18.o(.text);
